Render load indicator with an image on SSR
